Split files, better typing, better directory handling, small update to canvas api
